ai/voiceReport/requirements.txt

fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-multipart==0.0.9

numpy==1.26.4
pandas==2.2.2

torch==2.1.2

openai==1.84.0
requests==2.32.3
python-dotenv==1.0.1

pymysql